+++ Changelog.md
@@ -20,6 +20,8 @@ Compiler Features:
  * Standard JSON: New option ``modelCheckerSettings.targets`` allows specifying which targets should be checked. The valid options are ``all``, ``constantCondition``,
    ``underflow``, ``overflow``, ``divByZero``, ``balance``, ``assert``, ``popEmptyArray``, where the default is ``all``. Multiple targets can be chosen at the same time,
    separated by a comma without spaces: ``underflow,overflow,assert``.
+ * Command Line Interface: Allow "=" as separator between library name and address in ``--libraries`` commandline option.
+ * Command Line Interface: Only accept the library address that is prefixed with "0x" in ``--libraries`` commandline option.
 
 Bugfixes:
  * Code Generator: Fix length check when decoding malformed error data in catch clause.